feat(ocurrence-events-history): add audit
