+++ apps/backend/src/app/ocurrence-events-history/ocurrence-events-history.schema.ts
@@ -1,4 +1,7 @@
+
 import * as mongoose from 'mongoose';
+import { environment } from '../../environments/environment';
+import { AuditPlugin } from '@andes/mongoose-plugin-audit';
 
 export interface OcurrenceEventHistory {
     eventKey: string;
@@ -30,6 +33,9 @@ export const OcurrenceEventHistorySchema = new mongoose.Schema({
     originalRef: mongoose.SchemaTypes.ObjectId
 });
 
+if (environment.key) {
+    OcurrenceEventHistorySchema.plugin(AuditPlugin);
+}
 
 export const OcurrenceEventHistory = mongoose.model<OcurrenceEventHistoryDocument>(
     'ocurrence_event_history',
