+++ apps/backend/src/app/ocurrence-events/ocurrence-events.controller.ts
@@ -4,6 +4,8 @@ import { Request, Response } from '@andes/api-tool';
 import { authenticate, checkPermission } from '../application';
 import { InstitutionCtr } from '../institution/institution.controller'
 import * as mongoose from 'mongoose';
+import { OcurrenceEventHistory } from '../ocurrence-events-history/ocurrence-events-history.schema';
+import { OcurrenceEventHistoryCtr } from '../ocurrence-events-history/ocurrence-events-history.controller';
 
 class OcurrenceEventResource extends ResourceBase {
     Model = OcurrenceEvent;
@@ -21,6 +23,34 @@ class OcurrenceEventResource extends ResourceBase {
             }
             return next();
 
+        },
+        post: async (req: Request, res: Response, next) => {
+            const doc = req.body;
+            const eventHistory: OcurrenceEventHistory = {
+                eventKey: doc.eventKey,
+                activo: doc.activo,
+                fecha: doc.fecha,
+                institucion: doc.institucion,
+                indicadores: doc.indicadores,
+                originalRef: doc._id
+            };
+            const ocurrenceEventHistory = new OcurrenceEventHistory(eventHistory);
+            await OcurrenceEventHistoryCtr.create(ocurrenceEventHistory, req);
+            return next();
+        },
+        patch: async (req: Request, res: Response, next) => {
+            const doc = req.body;
+            const eventHistory: OcurrenceEventHistory = {
+                eventKey: doc.eventKey,
+                activo: doc.activo,
+                fecha: doc.fecha,
+                institucion: doc.institucion,
+                indicadores: doc.indicadores,
+                originalRef: doc._id
+            };
+            const ocurrenceEventHistory = new OcurrenceEventHistory(eventHistory);
+            await OcurrenceEventHistoryCtr.create(ocurrenceEventHistory, req);
+            return next();
         }
     };
     searchFileds = {
